requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.2.1
aiohttp==3.13.2
pandas==2.3.3
openpyxl==3.1.5
